Make adding series from UI possible
Add new form with name, value and time type selection.
The value and time type are currently printed via
Django template, as it was the simplest thing to do.

For better decoupling, the types should be loaded and
populated via AJAX, but for that the API needs a resource
for asking available types.